requirements.txt
pandas >= 1.2.0
numpy >= 1.20.0
scipy >= 1.6.0
tda-api
requests
lxml
python-dotenv
splinter
alpha_vantage
